refactor(mcp): split BaseChatMCP.swift into per-concern files (no behaviour change)
